HLT-138: enforce incremental SwiftLint quality
