+++ REUSE.toml
@@ -2,6 +2,7 @@ version = 1
 
 [[annotations]]
 path = [
+    ".swiftlint-baseline.json",
     ".claude/**",
     "Docs/**",
     "S2Y/Resources/**",
